+++ src/mediawiki.ts
@@ -779,6 +779,10 @@ class MediaWiki {
 	}
 
 	inTemplateArgument(expectName?: boolean): Tokenizer {
+		const regex = new RegExp(
+			`^(?:[^=|}{[<&~'_]|<(?!!--|(?:${Object.keys(this.config.tags).join('|')})[\\s/>]))*=`,
+			'iu',
+		);
 		return (stream, state) => {
 			const space = stream.eatSpace();
 			if (stream.eol()) {
@@ -790,16 +794,13 @@ class MediaWiki {
 			} else if (stream.match('}}')) {
 				state.tokenize = state.stack.pop()!;
 				return this.makeLocalTagStyle('templateBracket', state, 'nTemplate');
-			} else if (expectName && stream.eatWhile(/[^=|}{[<&~'_]/u)) {
-				if (stream.eat('=')) {
-					state.tokenize = this.inTemplateArgument();
-					return this.makeLocalTagStyle('templateArgumentName', state);
-				}
-				return this.makeLocalTagStyle('template', state);
-			} else if (stream.eatWhile(/[^|}{[<&~'_]/u) || space) {
-				return this.makeLocalTagStyle('template', state);
+			} else if (expectName && stream.match(regex)) {
+				state.tokenize = this.inTemplateArgument();
+				return this.makeLocalTagStyle('templateArgumentName', state);
 			}
-			return this.eatWikiText(modeConfig.tags.template)(stream, state);
+			return !isSolSyntax(stream) && stream.eatWhile(/[^|}{[<&~'_]/u) || space
+				? this.makeLocalTagStyle('template', state)
+				: this.eatWikiText(modeConfig.tags.template)(stream, state);
 		};
 	}